Fix broken Production build
There were some issues noted when running yarn build to create a production build images
These changes fixes the issues to allow successful production build to happen
