+++ Breakdown.md
@@ -1,0 +1,73 @@
+ The provided code snippet defines a set of HTTP endpoints using an asynchronous framework (likely inspired by frameworks like Express in Node.js). These endpoints handle various operations related to managing "SHL" objects, which likely represent some form of health link or configuration. Here's a summary of each CRUD method and what they do:
+
+### POST `/shl`
+- **Description**: Creates a new SHL (Health Link) based on the JSON payload provided in the request body.
+- **Steps**:
+  1. Extracts the JSON configuration from the request body.
+  2. Uses this configuration to create a new SHL entry in the database.
+  3. Responds with the newly created SHL object, excluding `files` and `config`.
+
+### POST `/shl/:shlId`
+- **Description**: Retrieves or updates an existing SHL based on its ID.
+- **Steps**:
+  1. Extracts a JSON payload containing configuration details for the SHL from the request body.
+  2. Validates and retrieves the SHL by its ID.
+  3. Checks if the SHL is active and not expired.
+  4. Verifies or updates the passcode if provided.
+  5. Generates a ticket for accessing the manifest.
+  6. Responds with file locations based on the SHL configuration, including both files and endpoints managed by the SHL.
+
+### GET `/shl/:shlId/file/:fileIndex`
+- **Description**: Retrieves a specific file from an existing SHL using its ID and file index.
+- **Steps**:
+  1. Extracts the ticket from the query parameters to verify access rights.
+  2. Validates the ticket against the SHL ID.
+  3. Retrieves the requested file from the database based on the provided indices.
+  4. Responds with the file content, setting appropriate headers for the content type.
+
+### GET `/shl/:shlId/endpoint/:endpointId`
+- **Description**: Retrieves a specific endpoint from an existing SHL using its ID and endpoint index.
+- **Steps**:
+  1. Extracts the ticket from the query parameters to verify access rights.
+  2. Validates the ticket against the SHL ID.
+  3. Retrieves the requested endpoint configuration and details from the database.
+  4. Encrypts and responds with the endpoint information, setting appropriate headers for the content type.
+
+### POST `/shl/:shlId/file`
+- **Description**: Adds a new file to an existing SHL using its ID.
+- **Steps**:
+  1. Extracts the management token from the authorization header.
+  2. Validates the management token and retrieves the managed SHL.
+  3. Parses the incoming file content and adds it to the database under the specified SHL.
+  4. Responds with the updated SHL details, including the newly added file.
+
+### POST `/shl/:shlId/endpoint`
+- **Description**: Adds a new endpoint to an existing SHL using its ID.
+- **Steps**:
+  1. Extracts the management token from the authorization header.
+  2. Validates the management token and retrieves the managed SHL.
+  3. Parses the incoming JSON configuration for the endpoint and adds it to the database under the specified SHL.
+  4. Responds with the updated SHL details, including the newly added endpoint.
+
+### PUT `/shl/:shlId`
+- **Description**: Updates an existing SHL using its ID with new configuration settings from the request body.
+- **Steps**:
+  1. Extracts the management token from the authorization header.
+  2. Validates the management token and retrieves the managed SHL.
+  3. Updates the SHL's configuration based on the provided JSON payload.
+  4. Responds with the updated SHL configuration details.
+
+### DELETE `/shl/:shlId`
+- **Description**: Deactivates an existing SHL using its ID.
+- **Steps**:
+  1. Extracts the management token from the authorization header.
+  2. Validates the management token and retrieves the managed SHL.
+  3. Deactivates the SHL in the database.
+  4. Responds with the deactivated status of the SHL.
+
+### POST `/subscribe`
+- **Description**: Subscribes to receive updates for a set of SHLs based on provided IDs and management tokens.
+- **Steps**:
+  1. Extracts an array of SHL IDs and management tokens from the request body.
+  2. Retrieves managed SHL objects for each ID and token pair.
+  3. Generates a subscription ticket and associates it with
